msreceipt/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
pika==1.1.0
psycopg2==2.9.1
gunicorn==20.1.0
django-heroku==0.3.1